+++ www/SQLitePlugin.js
@@ -447,7 +447,6 @@
         error: handlerFor(i, false)
       };
       tropts.push({
-        qid: null,
         sql: request.sql,
         params: request.params
       });