+++ scripts/beforePluginInstall.js
@@ -1,32 +1,26 @@
-#!/usr/bin/env node
-
 // Adapted from:
 // https://github.com/AllJoyn-Cordova/cordova-plugin-alljoyn/blob/master/scripts/beforePluginInstall.js
 
-var path = require('path');
-var exec = require('child_process').exec;
+const path = require('path');
+const exec = require('child_process').exec;
 
-// XXX TODO auto-detect:
-var package_name = 'cordova-sqlite-ext-core-common';
+const packageName = require('../package.json').name;
 
-module.exports = function (context) {
-    var Q = context.requireCordovaModule('q');
-    var deferral = new Q.defer();
+module.exports = function () {
+    return new Promise(function (resolve, reject) {
+        console.log('installing external dependencies via npm');
+        console.log('for package name: ' + packageName);
 
-    console.log('installing external dependencies via npm');
-
-    exec(   'npm install', {cwd: path.join('plugins', package_name)},
+        exec('npm install', { cwd: path.join('plugins', packageName) },
             function (error, stdout, stderr) {
                 if (error !== null) {
-                    // XXX TODO SIGNAL FAILURE HERE.
-                    console.log('npm install of external dependencies failed: ' + error);
-                    deferral.resolve();
+                    console.log('npm install of external dependencies failed with error message: ' + error.message);
+                    reject();
                 } else {
                     console.log('npm install of external dependencies ok');
-                    deferral.resolve();
+                    resolve();
                 }
             }
-    );
-
-    return deferral.promise;
+        );
+    });
 };