+++ scripts/beforePluginInstall.js
@@ -6,8 +6,8 @@
 var path = require('path');
 var exec = require('child_process').exec;
 
-// XXX FUTURE TBD auto-detect:
-var package_name = 'cordova-sqlite-storage';
+// XXX TODO auto-detect:
+var package_name = 'cordova-sqlite-ext-common-core';
 
 module.exports = function (context) {
     var Q = context.requireCordovaModule('q');