+++ spec/www/spec/db-tx-error-mapping-test.js
@@ -168,7 +168,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT with VALUES in the wrong place (with a trailing space) [XXX TBD "incomplete input" vs "syntax error" message on (WebKit) Web SQL on Android 8.x/...]', function(done) {
+        it(suiteName + 'INSERT with VALUES in the wrong place (with a trailing space) [XXX TBD "incomplete input" vs "syntax error" message on (WebKit) Web SQL on Android 8.x/XXX]', function(done) {
           var db = openDatabase("INSERT-Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -200,14 +200,18 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
+              /* ** XXX TODO NEEDS TO BE UPDATED:
               if (isWebSql && (/Android [7-9]/.test(navigator.userAgent)))
-                expect(error.message).toMatch(/could not prepare statement.*/); // XXX TBD incomplete input vs syntax error message on Android 8(+)
+                expect(error.message).toMatch(/could not prepare statement.*.../); // XXX TBD incomplete input vs syntax error message on Android 8(+)/XXX
               else if (isWebSql && !isChromeBrowser && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 near \"VALUES\": syntax error/);
               else if (isWebSql && isBrowser)
                 expect(error.message).toMatch(/could not prepare statement.*1 incomplete input/);
               else if (isWebSql)
                 expect(error.message).toMatch(/near \"VALUES\": syntax error/);
+              // XXX ... */
+              if (isWebSql)
+                expect(error.message).toMatch(/could not prepare statement.*/); // XXX TBD incomplete input vs syntax error message on Android 8(+) & iOS 12(+)
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (isAndroid && !isImpl2)
@@ -215,7 +219,8 @@ var mytests = function() {
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"VALUES\": syntax error.*code 1.*while compiling: INSERT INTO test_table/);
               else
-                expect(error.message).toMatch(/incomplete input/);
+                //* XXX ...
+                expect(error.message).toMatch(/incomplete input/); // XXX SQLite 3.22.0
 
               // FAIL transaction & check reported transaction error:
               return true;