+++ src/android/io/sqlc/SQLitePlugin.java
@@ -6,8 +6,6 @@
 
 package io.sqlc;
 
-import android.annotation.SuppressLint;
-
 import android.util.Log;
 
 import java.io.File;
@@ -30,7 +28,7 @@ import org.json.JSONObject;
 public class SQLitePlugin extends CordovaPlugin {
 
     /**
-     * Multiple database runner map (static).
+     * Concurrent database runner map.
      *
      * NOTE: no public static accessor to db (runner) map since it is not
      * expected to work properly with db threading.
@@ -44,7 +42,7 @@ public class SQLitePlugin extends CordovaPlugin {
      * THANKS to @NeoLSN (Jason Yang/楊朝傑) for giving the pointer in:
      * https://github.com/litehelpers/Cordova-sqlite-storage/issues/727
      */
-    static Map<String, DBRunner> dbrmap = new ConcurrentHashMap<String, DBRunner>();
+    private Map<String, DBRunner> dbrmap = new ConcurrentHashMap<String, DBRunner>();
 
     /**
      * NOTE: Using default constructor, no explicit constructor.