+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-###### cordova-sqlite-storage-ext-core-common 1.0.0
+###### cordova-sqlite-storage-ext-core-common 1.0.1-dev
 
 - Use SQLite 3.26.0 (with a security update and support for window functions) with SQLITE_DEFAULT_SYNCHRONOUS=3 (EXTRA DURABLE) from cordova-sqlite-storage-dependencies 1.2.1 in this plugin version branch