+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-###### cordova-sqlite-storage-ext-core-common 1.0.1-dev
+###### cordova-sqlite-storage-ext-core-common 2.0.0
 
 - SQLITE_DBCONFIG_DEFENSIVE flag - iOS/macOS/Windows (POTENTIALLY BREAKING CHANGE)
 - remove internal qid usage from JavaScript (not needed)