+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+###### cordova-sqlite-ext-core-common 0.1.0-dev
+
 ###### cordova-sqlite-storage-ext-core-common 2.0.0
 
 - SQLITE_DBCONFIG_DEFENSIVE flag - iOS/macOS/Windows (POTENTIALLY BREAKING CHANGE)