+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-##### cordova-sqlite-ext-common-core 0.1.0-dev
+##### cordova-sqlite-ext-common-core 0.1.0
 
 ###### cordova-sqlite-ext-core-common 0.1.0