+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-##### cordova-sqlite-ext-common-core 0.1.1-dev
+###### cordova-sqlite-extcore 0.1.0
 
 - move the embedded `SQLite3-WinRT` component to `src/windows/SQLite3-WinRT-sync` and update `plugin.xml`